implement getDashboards (#26481)
* implement getInsightSubjects

NOTE: Only partial implementation. Still need to make actual API call.

* implement getDashboards

* add error checks

for properties missing from GraphQL API

* fix error when saving settings insight

* remove unused dashboard hook

+++ client/web/src/enterprise/insights/pages/dashboards/dashboard-page/components/delete-dashboard-modal/hooks/use-delete-dashboard-handler.ts
@@ -28,6 +28,10 @@ export function useDeleteDashboardHandler(props: UseDeleteDashboardHandlerProps)
         setLoadingOrError(true)
 
         try {
+            if (!dashboard.settingsKey || !dashboard.owner) {
+                throw new Error('TODO: implement deleteDashboard for GraphQl API')
+            }
+
             await deleteDashboard({
                 dashboardSettingKey: dashboard.settingsKey,
                 dashboardOwnerId: dashboard.owner.id,
